support archetype from local-catalog
