+++ src/archetype/Archetype.ts
