+++ src/model/XmlSchema.ts
@@ -1,45 +1,45 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license.
 
-export interface IPomRoot {
-    project?: IPomProject;
-}
+// export interface IPomRoot {
+//     project?: IPomProject;
+// }
 
-export interface IPomProject {
-    modules?: IPomModules[];
-    artifactId?: IPomArtifactId[];
-}
+// export interface IPomProject {
+//     modules?: IPomModules[];
+//     artifactId?: IPomArtifactId[];
+// }
 
-export interface IPomModules {
-    // tslint:disable-next-line:no-reserved-keywords
-    module?: IPomModule[];
-}
+// export interface IPomModules {
+//     // tslint:disable-next-line:no-reserved-keywords
+//     module?: IPomModule[];
+// }
 
-export type IPomModule = string[];
-export type IPomArtifactId = string[];
+// export type IPomModule = string[];
+// export type IPomArtifactId = string[];
 
-export interface IArchetypeCatalogRoot {
-    "archetype-catalog"?: IArchetypeCatalog;
-}
+// export interface IArchetypeCatalogRoot {
+//     "archetype-catalog"?: IArchetypeCatalog;
+// }
 
-export interface IArchetypeCatalog {
-    archetypes?: IArchetypes[];
-}
+// export interface IArchetypeCatalog {
+//     archetypes?: IArchetypes[];
+// }
 
-export interface IArchetypes {
-    archetype?: IArchetype[];
-}
+// export interface IArchetypes {
+//     archetype?: IArchetype[];
+// }
 
-export interface IArchetype {
-    groupId?: IArchetypeGroupId[];
-    artifactId?: IArchetypeArtifactId[];
-    version?: IArchetypeVersion[];
-    description?: IArchetypeDescription[];
-    repository?: IArchetypeRepository[];
-}
+// export interface IArchetype {
+//     groupId?: IArchetypeGroupId[];
+//     artifactId?: IArchetypeArtifactId[];
+//     version?: IArchetypeVersion[];
+//     description?: IArchetypeDescription[];
+//     repository?: IArchetypeRepository[];
+// }
 
-export type IArchetypeGroupId = string[];
-export type IArchetypeArtifactId = string[];
-export type IArchetypeVersion = string[];
-export type IArchetypeDescription = string[];
-export type IArchetypeRepository = string[];
+// export type IArchetypeGroupId = string[];
+// export type IArchetypeArtifactId = string[];
+// export type IArchetypeVersion = string[];
+// export type IArchetypeDescription = string[];
+// export type IArchetypeRepository = string[];
